requirements.txt
selenium
requests
beautifulsoup4
python-dotenv